backend/requirements.txt
# Core
python-dotenv
httpx
numpy
scipy  # For benchmark optimizers (Markowitz, HRP)

# D-Wave Ocean SDK (QUBO solver + simulated annealing)
dimod
dwave-neal

# LangGraph (agent orchestration)
langgraph>=0.2.0
langchain-core>=0.3.0

# AWS Braket (quantum RNG — optional, only for real QPU)
# amazon-braket-sdk
# boto3

# Sui integration (used by sui_client.py + ptb_builder.py)
pysui

# API server (relayer_server.py + event_provider.py)
fastapi
uvicorn[standard]
websockets
pydantic

# Testing
pytest
